Kickstart app onboarding flow (#58)
Add a rough onboarding flow for new pills.

This flow waits for the first sample data to be received by any pill that is not in the database already, saves a new pill along with the dataset in the db and selects it as the current device, before navigating back to the previous screen.